// management/frontend/src/api/zod/workspace-business-lines/workspace-business-lines.ts
export const workspaceBusinessLinesReadBusinessLinesQuerySkipDefault = 0;
export const workspaceBusinessLinesReadBusinessLinesQuerySkipMin = 0;

export const workspaceBusinessLinesReadBusinessLinesQueryLimitDefault = 20;
export const workspaceBusinessLinesReadBusinessLinesQueryLimitMax = 500;



export const WorkspaceBusinessLinesReadBusinessLinesQueryParams = zod.object({
  "skip": zod.number().min(workspaceBusinessLinesReadBusinessLinesQuerySkipMin).default(workspaceBusinessLinesReadBusinessLinesQuerySkipDefault),
  "limit": zod.number().max(workspaceBusinessLinesReadBusinessLinesQueryLimitMax).default(workspaceBusinessLinesReadBusinessLinesQueryLimitDefault),
  "name": zod.union([zod.string(),zod.null()]).optional(),
  "status": zod.union([zod.number(),zod.null()]).optional()
})